Added null checking for TransactionalMap method args
Also some testing cleanup.
