+++ hazelcast/src/main/java/com/hazelcast/core/TransactionalMap.java
@@ -76,6 +76,7 @@ public interface TransactionalMap<K, V> extends TransactionalObject, BaseMap<K, 
      * Transactional implementation of {@link IMap#containsKey(Object)}.
      *
      * @see IMap#containsKey(Object)
+     * @throws NullPointerException if the specified key is null.
      */
     boolean containsKey(Object key);
 
@@ -83,6 +84,7 @@ public interface TransactionalMap<K, V> extends TransactionalObject, BaseMap<K, 
      * Transactional implementation of {@link IMap#get(Object)}.
      *
      * @see IMap#get(Object)
+     * @throws NullPointerException if the specified key is null.
      */
     V get(Object key);
 
@@ -91,6 +93,7 @@ public interface TransactionalMap<K, V> extends TransactionalObject, BaseMap<K, 
      * Lock will be released at the end of the transaction (either commit or rollback).
      *
      * @see IMap#get(Object)
+     * @throws NullPointerException if the specified key is null.
      */
     V getForUpdate(Object key);
 
@@ -114,6 +117,7 @@ public interface TransactionalMap<K, V> extends TransactionalObject, BaseMap<K, 
      * The object to be put will be accessible only in the current transaction context till transaction is committed.
      *
      * @see IMap#put(Object, Object)
+     * @throws NullPointerException if the specified key or value is null.
      */
     V put(K key, V value);
 
@@ -123,6 +127,7 @@ public interface TransactionalMap<K, V> extends TransactionalObject, BaseMap<K, 
      * The object to be put will be accessible only in the current transaction context till transaction is committed.
      *
      * @see IMap#put(Object, Object, long, java.util.concurrent.TimeUnit)
+     * @throws NullPointerException if the specified key, value or timeunit is null.
      */
     V put(K key, V value, long ttl, TimeUnit timeunit);
 
@@ -132,6 +137,7 @@ public interface TransactionalMap<K, V> extends TransactionalObject, BaseMap<K, 
      * The object to be set will be accessible only in the current transaction context till transaction is committed.
      *
      * @see IMap#set(Object, Object)
+     * @throws NullPointerException if the specified key or value is null.
      */
     void set(K key, V value);
 
@@ -141,6 +147,7 @@ public interface TransactionalMap<K, V> extends TransactionalObject, BaseMap<K, 
      * The object to be put will be accessible only in the current transaction context until the transaction is committed.
      *
      * @see IMap#putIfAbsent(Object, Object)
+     * @throws NullPointerException if the specified key or value is null.
      */
     V putIfAbsent(K key, V value);
 
@@ -150,6 +157,7 @@ public interface TransactionalMap<K, V> extends TransactionalObject, BaseMap<K, 
      * The object to be replaced will be accessible only in the current transaction context until the transaction is committed.
      *
      * @see IMap#replace(Object, Object)
+     * @throws NullPointerException if the specified key or null.
      */
     V replace(K key, V value);
 
@@ -159,6 +167,7 @@ public interface TransactionalMap<K, V> extends TransactionalObject, BaseMap<K, 
      * The object to be replaced will be accessible only in the current transaction context until the transaction is committed.
      *
      * @see IMap#replace(Object, Object, Object)
+     * @throws NullPointerException if the specified key, oldValue or newValue is null.
      */
     boolean replace(K key, V oldValue, V newValue);
 
@@ -168,6 +177,7 @@ public interface TransactionalMap<K, V> extends TransactionalObject, BaseMap<K, 
      * The object to be removed will be removed from only the current transaction context until the transaction is committed.
      *
      * @see IMap#remove(Object)
+     * @throws NullPointerException if the specified key is null.
      */
     V remove(Object key);
 
@@ -177,6 +187,7 @@ public interface TransactionalMap<K, V> extends TransactionalObject, BaseMap<K, 
      * The object to be deleted will be removed from only the current transaction context until the transaction is committed.
      *
      * @see IMap#delete(Object)
+     * @throws NullPointerException if the specified key is null.
      */
     void delete(Object key);
 
@@ -186,6 +197,7 @@ public interface TransactionalMap<K, V> extends TransactionalObject, BaseMap<K, 
      * The object to be removed will be removed from only the current transaction context until the transaction is committed.
      *
      * @see IMap#remove(Object, Object)
+     * @throws NullPointerException if the specified key or value null.
      */
     boolean remove(Object key, Object value);
 
@@ -200,6 +212,7 @@ public interface TransactionalMap<K, V> extends TransactionalObject, BaseMap<K, 
      * Transactional implementation of {@link IMap#keySet(com.hazelcast.query.Predicate)}.
      *
      * @see IMap#keySet(com.hazelcast.query.Predicate)
+     * @throws NullPointerException if the specified predicate is null.
      */
     Set<K> keySet(Predicate predicate);
 
@@ -214,6 +227,7 @@ public interface TransactionalMap<K, V> extends TransactionalObject, BaseMap<K, 
      * Transactional implementation of {@link IMap#values(com.hazelcast.query.Predicate)}.
      *
      * @see IMap#values(com.hazelcast.query.Predicate)
+     * @throws NullPointerException if the specified predicate is null.
      */
     Collection<V> values(Predicate predicate);
 }
