+++ hazelcast/src/test/java/com/hazelcast/map/impl/tx/MapTransactionLockingTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.map;
+package com.hazelcast.map.impl.tx;
 
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.core.IMap;
