+++ hazelcast/src/test/java/com/hazelcast/multimap/MultiMapTransactionStressTest.java
@@ -23,7 +23,7 @@ import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.core.IQueue;
 import com.hazelcast.core.MultiMap;
 import com.hazelcast.core.TransactionalMultiMap;
-import com.hazelcast.map.MapTransactionStressTest;
+import com.hazelcast.map.impl.tx.MapTransactionStressTest;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.HazelcastTestSupport;
 import com.hazelcast.test.annotation.NightlyTest;
